node.scm (rrepl-eval): Bugfix: Use 'write' instead of 'write-line'
The 'rrepl-eval' is always failing to send sexps that contain strings in the
proper way because the procedure uses the human-readable printing to transmit
the data.  As the result an RREPL fails to evaluate it and reports an error.
This patch fixes that bug by switching to the machine-readable data printing.

Reported by Ludovic Courtès, in
<http://permalink.gmane.org/gmane.lisp.guile.user/11974>

* modules/ssh/dist/node.scm (rrepl-eval): Bugfix: Use 'write' instead of
  'write-line' to send the code for evaluation to a RREPL so it will be
  machine-readable.