node.scm (node-open-rrepl): Start an RREPL automatically
Start an RREPL automatically if it is not running on the remote side.  The
user can turn this feature off by means of a new 'start-repl-server?' argument
accepted by 'make-node' procedure.  The code uses 'pgrep' program to check
whether an RREPL is running or not on the node.

Sugguested by Ludovic Courtès, in
<http://permalink.gmane.org/gmane.lisp.guile.user/11974>

* modules/ssh/dist/node.scm (node-open-rrepl): Start an RREPL automatically if
  there is no RREPL running on the node.
  (make-node): Accept 'start-repl-server?' keyworded argument that specifies
  whether an RREPL should be run on the start or not.
  (rrepl-running?, rexec): New procedures.