tests/tunnel.scm ("call-with-ssh-forward"): Free allocated resources
The test suite failes due to an overlap between "port forwarding, direct" and
"call-with-ssh-forward".  Presumably a process that spawned by "port
forwarding, direct" keeps running and messes things up in the later tests.
The solution seems to be that all the resources allocated by the "port
forwarding, direct" should be freed as soon as they don't needed.

* tests/tunnel.scm ("call-with-ssh-forward"): Free allocated resources at the
  end of the test.