node.scm (rrepl-get-result): Bugfix: Handle compilation errors
Guile-SSH would always fail to read compilation errors properly because it
considered the message as "undefined" result.  Now this bug should be fixed.

An example of an error that now should be handled is "no code for module" due
to using a non-existing module in 'with-ssh' expression.

Reported by Mathieu, in
<https://github.com/artyom-poptsov/guile-ssh/issues/3>

* modules/ssh/dist/node.scm (rrepl-get-result): Bugfix: Do not consider
  compilation errors as undefined results.
  (rrepl-eval): Call 'exit' on the remote side to ensure that the remote side
  closed a channel.
* tests/dist.scm: Improve logging.
  ("rrepl-get-result, compilation error"): New TC.
* doc/guile-ssh.texi (Acknowledgments): Update.
* AUTHORS, NEWS, THANKS: Update.