libguile-ssh/channel-type.c: Bugfix
When a remove side closing a channel Guile-SSH would fail as the libssh
callback that marks the Guile-SSH channel as closed tried to convert a Guile
channel object to the channel data and fails to do it.

Now the problem should be fixed.

Reported by Ludovic Courtès <ludo@gnu.org> in an email, here:
<https://debbugs.gnu.org/cgi/bugreport.cgi?bug=52975>

* libguile-ssh/channel-type.c (ssh_channel_to_scm): Don't pass an SCM object
  to the libssh callbacks as it won't make it through the libssh internals; pass
  a raw channel data instead.
  (channel_close_callback): Convert a userdata to Guile-SSH channel data
  immediately.